The Structured Wright Fisher Model for IBD

Simulate a population forwards with recombination that approximates the Structured Wright Fisher Process and tracks haplotype identity by descent where individuals represent demes, such that within a deme individual-level COI is considered.

sim_swf(pos, N, m, rho, mean_coi, tlim, migr_dist_mat = NULL, verbose = FALSE)

Arguments

pos
vector; the genomic coordinates for chromosome and position of the sites

N
numeric numeric; The number of individuals to consider in each deme

m
numeric numeric; Probability of internal migration where m represents the probability of moving from host_origin to host_new by m*(1-1/N) of each deme

rho
numeric; expected recombination rate

mean_coi
numeric vector; The lambda of a right-shifted Poisson process, 1 + Pos(lambda) representing the average COI of each deme

tlim
numeric; the maximum number of generations to consider before exiting gracefully if all samples have not coalesced

migr_dist_mat
numeric matrix; Migrations rates with origin specified as rows and destination in columns.

verbose
boolean

Value

Returns a list of length six that contains the COI of each individual. A recombination list of length of tlim where each element contains the recombination block -- as a boolean -- of the two parental haplotypes. (the number of generations it took for all lineages to coalesce). Finally, there are lists for the parental host and parental haplotype assignments for the "paternal" and "maternal" haplotypes (1 and 2), respectively.

Details

Demes are assumed to be ordered throughout (i.e. the order needs to be consistent between N, m, mean_coi, and the rows and columns of the migration matrix).

Migration matrix is assumed to be a distance matrix, where the user can specify the scalar through the dist_scalar parameter